refactor: split schemas into the separate files
